\name{JointScoreTest-package}
\alias{JointScoreTest-package}
\alias{JointScoreTest}
\docType{package}
\title{
\packageTitle{JointScoreTest}
}
\description{
This package implements the score test for jointly testing the fixed and random effects in generalized linear mixed models (GLMMs).
}
\details{
\tabular{ll}{
Package: \tab JointScoreTest\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2018-06-08\cr
License: \tab GPL (>= 2)\cr
}
This package implements the score test for jointly testing the fixed and random effects in generalized linear mixed models (GLMMs). In order to get p-value of the score test, the user should provide the design matrix for fixed effects, 
random effects and the correlation matrix for the random effects in the `JointScoreTest’ function.
}
\author{
Jingchunzi Shi, Seunggeun Lee

Maintainer: Jingchunzi (Jing) Shi <shijingc@umich.edu>

}
\keyword{ Joint test; Fixed effects; Random effects; GLMM. }

